merge from master

+++ app/Http/Controllers/Manager/Role/RoleController.php
@@ -18,13 +18,15 @@ class RoleController extends Controller
      */
     public function index()
     {
-        $this->authorize('role.view');
+        $this->authorize('manager.role.view');
 
         return view('dashboard.role.list', ['roles' => Role::all()]);
     }
 
     public function create()
     {
+        $this->authorize('manager.role.create');
+
         return view('dashboard.role.create', [
             'groups' => PermissionGroup::with('permissions')->get()
         ]);
@@ -32,6 +34,8 @@ class RoleController extends Controller
 
     public function view(string $id, Role $role)
     {
+        $this->authorize('manager.role.view');
+
         return view('dashboard.role.view', [
             'role' => $role->with('permissions')->findOrFail($id),
             'groups' => PermissionGroup::with('permissions')->get()
@@ -40,13 +44,14 @@ class RoleController extends Controller
 
     public function update(string $id, UpdateRole $request)
     {
-        $role = Role::findOrFail($id)->fill($request->only('name'));
+        $role = Role::findOrFail($id);
 
-        if ($request->permissions) {
-            $role->syncPermissions($this->permissionIdToName($request->permissions));
-        }
+        $role->forceFill([
+            'name'     => $request->name,
+            'customer' => (bool) $request->for_customer
+        ])->save();
 
-        $role->forceFill(['customer' => (bool) $request->for_customer])->save();
+        $role->permissions()->sync($request->permissions ?? []);
 
         return back()->with('success', 'Cập nhật thành công');
     }
@@ -70,6 +75,8 @@ class RoleController extends Controller
 
     public function delete(string $id, Role $role)
     {
+        $this->authorize('manager.role.delete');
+
         $role->findOrFail($id)->delete();
 
         return redirect(route('manager.role'))->with('success', 'Xóa thành công');
